requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic==2.9.0
pandas==2.2.0
scikit-learn==1.5.0
joblib==1.4.0
python-dotenv==1.0.0
